+++ tests/testing/test_ldap.py
@@ -10,24 +10,10 @@ if TYPE_CHECKING:  # pragma: no cover
 
 
 @pytest.mark.integration
-def test_get_merged_person_from_login_success(
-    testing_app_client_that_is_ldap_authenticated: TestClient,
-) -> None:
-    response = testing_app_client_that_is_ldap_authenticated.post(
-        "/v0/merged-person-from-login"
-    )
-    assert response.status_code == status.HTTP_200_OK, response.text
-    data = response.json()
-    assert data["email"] == ["Writer@rki.com"]
-    assert data["fullName"] == ["Writer"]
-    assert data["$type"] == "MergedPerson"
-
-
-@pytest.mark.integration
 def test_search_persons_or_contact_points_in_ldap_success(
-    testing_app_client_with_api_key_read_permission: TestClient,
+    testing_app_client_with_bearer_write_permission: TestClient,
 ) -> None:
-    response = testing_app_client_with_api_key_read_permission.get(
+    response = testing_app_client_with_bearer_write_permission.get(
         "/v0/ldap", params={"q": "mex", "limit": 1}
     )
     assert response.status_code == status.HTTP_200_OK, response.text